An example of how this might be simplified
