+++ example/app.js
@@ -1,0 +1,11 @@
+const routering = require('../lib/index')
+const onUrlChange = require('on-url-change');
+
+module.exports = function(mutate){
+  const state = {}
+
+  const emitter = onUrlChange();
+  emitter.on('change', () => mutate.set(state, 'route', window.location.pathname))
+
+  return state
+}
