+++ lib/index.js
@@ -1,37 +1,19 @@
-module.exports = function (opts, setRoute) {
-  document.onclick = function (e) {
-    if (e.target.href) {
-      const href = e.target.getAttribute('href')
-      if (href && !href.match(/.*?\/\//)) {
-        changeRoute(href)()
-        e.preventDefault()
-      }
+function setPath(path){
+    if(path !== window.location.pathname){
+      window.history.pushState({ path }, document.title, path)
     }
-  }
+}
 
-  if (window.location.pathname === '/' && opts.defaultRoute !== '/') {
-    setRoute(opts.defaultRoute)
-    window.history.pushState({ route: opts.defaultRoute }, document.title, opts.defaultRoute)
-  }
+function pushStateAnchors(event) {
+    const href = event.target.getAttribute('href')
 
-  function changeRoute (newRoute) {
-    return function (event) {
-      event && event.preventDefault()
-      let route = typeof newRoute === 'string' ? newRoute : newRoute()
-      setRoute(route)
-      if (route === '/' && opts.defaultRoute !== '/') {
-        route = opts.defaultRoute
-      }
-      setRoute(route)
-      window.history.pushState({ route }, document.title, route)
+    if (href && !href.match(/.*?\/\//)) {
+      event.preventDefault()
+      setPath(href)
     }
-  }
+}
 
-  window.addEventListener('popstate', (event) => {
-    if (event.state.route) {
-      setRoute(event.state.route)
-    }
-  })
-
-  return changeRoute
+module.exports = {
+  setPath,
+  pushStateAnchors
 }
